avoid using size_t

+++ src/vlGraphics/DoubleVertexRemover.cpp
@@ -36,7 +36,7 @@ using namespace vl;
 
 //-----------------------------------------------------------------------------
 template<class T>
-ref<ArrayAbstract> VertexMapper::regenerateT(ArrayAbstract* data, const std::vector<size_t>& map_new_to_old) const
+ref<ArrayAbstract> VertexMapper::regenerateT(ArrayAbstract* data, const std::vector<u32>& map_new_to_old) const
 {
   ref<T> in_data = cast<T>(data);
   if (in_data)
@@ -50,7 +50,7 @@ ref<ArrayAbstract> VertexMapper::regenerateT(ArrayAbstract* data, const std::vec
   return NULL;
 }
 //-----------------------------------------------------------------------------
-ref<ArrayAbstract> VertexMapper::regenerate(ArrayAbstract* data, const std::vector<size_t>& map_new_to_old) const
+ref<ArrayAbstract> VertexMapper::regenerate(ArrayAbstract* data, const std::vector<u32>& map_new_to_old) const
 {
   ref<ArrayAbstract> out_data;
 
@@ -158,17 +158,17 @@ void DoubleVertexRemover::removeDoubles(Geometry* geom)
   mMapNewToOld.clear();
   mMapOldToNew.clear();
 
-  size_t vert_count = geom->vertexArray() ? geom->vertexArray()->size() : geom->vertexAttribArray(VA_Position) ? geom->vertexAttribArray(VA_Position)->data()->size() : 0;
+  u32 vert_count = (u32)(geom->vertexArray() ? geom->vertexArray()->size() : geom->vertexAttribArray(VA_Position) ? geom->vertexAttribArray(VA_Position)->data()->size() : 0);
   
   VL_CHECK(vert_count);
   if (!vert_count)
     return;
 
-  std::vector<size_t> verti;
+  std::vector<u32> verti;
   verti.resize(vert_count);
   mMapOldToNew.resize(vert_count);
 
-  for(unsigned int i=0; i<verti.size(); ++i)
+  for(u32 i=0; i<verti.size(); ++i)
   {
     verti[i] = i;
     mMapOldToNew[i] = 0xFFFFFFFF;
@@ -177,20 +177,20 @@ void DoubleVertexRemover::removeDoubles(Geometry* geom)
   std::sort(verti.begin(), verti.end(), LessCompare(geom));
   EqualsCompare equal_vertex(geom);
   mMapNewToOld.reserve(vert_count);
-  unsigned int unique_vert_idx = 0;
+  u32 unique_vert_idx = 0;
   for(unsigned i=1; i<verti.size(); ++i)
   {
     if ( !equal_vertex(verti[unique_vert_idx], verti[i]) )
     {
       for(unsigned j=unique_vert_idx; j<i; ++j)
-        mMapOldToNew[verti[j]] = mMapNewToOld.size();
+        mMapOldToNew[verti[j]] = (u32)mMapNewToOld.size();
       mMapNewToOld.push_back(verti[unique_vert_idx]);
       unique_vert_idx = i;
     }
   }
   for(unsigned j=unique_vert_idx; j<verti.size(); ++j)
   {
-    mMapOldToNew[verti[j]] = mMapNewToOld.size();
+    mMapOldToNew[verti[j]] = (u32)mMapNewToOld.size();
     mMapNewToOld.push_back(verti[unique_vert_idx]);
   }
 
@@ -205,13 +205,13 @@ void DoubleVertexRemover::removeDoubles(Geometry* geom)
     draw_cmd.push_back( geom->drawCalls()->at(idraw) );
   geom->drawCalls()->clear();
 
-  for(size_t idraw=0; idraw<draw_cmd.size(); ++idraw)
+  for(u32 idraw=0; idraw<draw_cmd.size(); ++idraw)
   {
     ref<DrawElementsUInt> de = new DrawElementsUInt( draw_cmd[idraw]->primitiveType() );
     geom->drawCalls()->push_back(de.get());
-    const int idx_count = draw_cmd[idraw]->countIndices();
+    const u32 idx_count = draw_cmd[idraw]->countIndices();
     de->indexBuffer()->resize(idx_count);
-    int i=0;
+    u32 i=0;
     for(IndexIterator it = draw_cmd[idraw]->indexIterator(); it.hasNext(); it.next(), ++i)
       de->indexBuffer()->at(i) = mMapOldToNew[it.index()];
   }
